requirements.txt
selenium
notify-send
